updated photography link and album
The Photography section of my website previously linked to a journal where a few of my photos had been published, but it was an old edition and not relevant anymore. I updated it to link to an album on Google Photos that I can add to - might change it to something nicer eventually. Also updated the link image to match.